Alignment merge from trunk from revision difference 542:549
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/branches/FermaNext-qt4-fullport@559 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ src/TrussUnit/TrussUnitWindow.cpp
@@ -525,17 +525,19 @@ bool TrussUnitWindow::inBDiagResizeRect ( int x, int y )
     if ( maximized )
         return false;
 
-    if ( sqrt( (x - (windowRightBottomPos.x() - bordWidth)) * 
-        (x - (windowRightBottomPos.x() - bordWidth)) +
-        (y - (windowLeftTopPos.y() + bordWidth)) * 
-        (y - (windowLeftTopPos.y() + bordWidth)) ) <= resEllRad )
+    if ( sqrt( double((x - (windowRightBottomPos.x() - bordWidth)) * 
+                      (x - (windowRightBottomPos.x() - bordWidth)) +
+                      (y - (windowLeftTopPos.y() + bordWidth)) * 
+                      (y - (windowLeftTopPos.y() + bordWidth)) )) 
+         <= resEllRad )
     {
         return true;
     }
-    else if ( sqrt( (x - (windowLeftTopPos.x() + bordWidth)) * 
-            (x - (windowLeftTopPos.x() + bordWidth)) + 
-            (y - (windowRightBottomPos.y() - bordWidth)) * 
-            (y - (windowRightBottomPos.y() - bordWidth)) ) <= resEllRad )
+    else if ( sqrt( double((x - (windowLeftTopPos.x() + bordWidth)) * 
+                           (x - (windowLeftTopPos.x() + bordWidth)) + 
+                           (y - (windowRightBottomPos.y() - bordWidth)) * 
+                           (y - (windowRightBottomPos.y() - bordWidth))) ) 
+              <= resEllRad )
     {
         return true;
     }
@@ -547,17 +549,19 @@ bool TrussUnitWindow::inFDiagResizeRect ( int x, int y )
     if ( maximized )
         return false;
 
-    if ( sqrt( (x - (windowLeftTopPos.x() + bordWidth)) * 
-        (x - (windowLeftTopPos.x() + bordWidth)) +
-        (y - (windowLeftTopPos.y() + bordWidth)) * 
-        (y - (windowLeftTopPos.y() + bordWidth)) ) <= resEllRad )
+    if ( sqrt( double((x - (windowLeftTopPos.x() + bordWidth)) * 
+                      (x - (windowLeftTopPos.x() + bordWidth)) +
+                      (y - (windowLeftTopPos.y() + bordWidth)) * 
+                      (y - (windowLeftTopPos.y() + bordWidth))) ) 
+         <= resEllRad )
     {
         return true;
     }
-    else if ( sqrt( (x - (windowRightBottomPos.x() - bordWidth)) * 
-            (x - (windowRightBottomPos.x() - bordWidth)) +
-            (y - (windowRightBottomPos.y() - bordWidth)) * 
-            (y - (windowRightBottomPos.y() - bordWidth)) ) <= resEllRad )
+    else if ( sqrt( double((x - (windowRightBottomPos.x() - bordWidth)) * 
+                           (x - (windowRightBottomPos.x() - bordWidth)) +
+                           (y - (windowRightBottomPos.y() - bordWidth)) * 
+                           (y - (windowRightBottomPos.y() - bordWidth))) ) 
+              <= resEllRad )
     {
         return true;
     }
